requirements-dev.txt
Cython>=3.0.0
pytest
build
twine
python-memtrace